+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "cSpell.words": ["mikro"]
+  "cSpell.words": ["chatable", "mikro"]
 }
