+++ src/feature/api/channels/infrastructure/channel-banned-user.entity.ts
