+++ src/feature/api/channels/infrastructure/channel.entity.ts
@@ -8,7 +8,7 @@ import {
 } from '@mikro-orm/core';
 import { v4 } from 'uuid';
 import { ChannelMessageEntity } from './channelMessage.entity';
-import { ChannelBannedUserEntity } from './channelBannedUser.entity';
+import { ChannelBannedUserEntity } from './channel-banned-user.entity';
 import { ChannelParticipantEntity } from './channelParticipant.entity';
 
 @Entity({ tableName: 'channel' })
