+++ src/feature/api/channels/presentation/gateway/channel-message.repository.ts
@@ -1,6 +1,6 @@
 import { EntityManager } from '@mikro-orm/postgresql';
 import { Injectable } from '@nestjs/common';
-import { ChannelMessageEntity } from 'src/feature/api/channel/infrastructure/channelMessage.entity';
+import { ChannelMessageEntity } from 'src/feature/api/channels/infrastructure/channelMessage.entity';
 
 @Injectable()
 export class ChannelMessageRepository {
