+++ src/feature/api/channels/presentation/gateway/dto/my-channel.dto.ts
