+++ src/feature/api/channels/presentation/gateway/dto/public-channel.dto.ts
