+++ src/feature/api/users/infrastructure/banned-user.entity.ts
