--- src/feature/gateway/gateway.module.ts
@@ -1,7 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ChannelModule } from './channel/channel.module';
-
-@Module({
-  imports: [ChannelModule],
-})
-export class GatewayModule {}
